JavaFX is a rather problematic UI framework... - Encapsulated JavaFX madness into separate View-Classes - Completely overhauled Communication to the JavaFX Controllers - Commented out checks and UI for Overview to check it works at all
see #24
