+++ src/main/java/de/vogel612/helper/TranslationHelper.java
@@ -3,10 +3,11 @@ package de.vogel612.helper;
 import de.vogel612.helper.data.OverviewModel;
 import de.vogel612.helper.ui.OverviewPresenter;
 import de.vogel612.helper.ui.OverviewView;
+import de.vogel612.helper.ui.ResxChooser;
 import de.vogel612.helper.ui.jfx.JFXOverviewView;
-import de.vogel612.helper.ui.jfx.JFXResxChooser;
-import de.vogel612.helper.ui.swing.SwingOverviewView;
+import de.vogel612.helper.ui.jfx.JFXResxChooserController;
 import de.vogel612.helper.ui.TranslationPresenter;
+import de.vogel612.helper.ui.jfx.JFXResxChooserView;
 
 import java.nio.file.Path;
 import java.nio.file.Paths;
@@ -15,7 +16,6 @@ import javafx.application.Application;
 import javafx.fxml.FXMLLoader;
 import javafx.scene.Scene;
 import javafx.scene.layout.GridPane;
-import javafx.scene.layout.StackPane;
 import javafx.stage.Stage;
 import javafx.stage.StageStyle;
 
@@ -36,11 +36,7 @@ public class TranslationHelper extends Application {
 
     @Override
     public void start(Stage primaryStage) throws Exception {
-        // FIXME: Make this the DI Context root and wire up stuff here instead of in the OverviewPresenter!
-        FXMLLoader resxLoader = new FXMLLoader(getClass().getResource("/ResxChooser.fxml"));
-
-        GridPane rcPane = resxLoader.load();
-        JFXResxChooser rc = resxLoader.getController();
+        ResxChooser rc = new JFXResxChooserView(primaryStage, getClass().getResource("/ResxChooser.fxml"));
         Parameters params = getParameters();
         if (params.getUnnamed().size() != 0) { // should be 1..
             final Path resxFile = Paths.get(params.getUnnamed().get(0));
@@ -49,14 +45,10 @@ public class TranslationHelper extends Application {
 
         // That's a view though.. duh.
         TranslationPresenter tp = new TranslationPresenter();
-
-        FXMLLoader overviewLoader = new FXMLLoader(getClass().getResource("/OverviewView.fxml"));
-        StackPane overviewPane = overviewLoader.load();
-        JFXOverviewView v = overviewLoader.getController();
+        OverviewView v = new JFXOverviewView(primaryStage, getClass().getResource("/OverviewView.fxml"));
 
         OverviewModel m = new OverviewModel();
-//        OverviewPresenter p = new OverviewPresenter(primaryStage, rcPane, overviewPane);
-        OverviewPresenter p = new OverviewPresenter(null, null, null, null); // just to stop the compiler from screaming at me
+        OverviewPresenter p = new OverviewPresenter(m, v, tp, rc); // just to stop the compiler from screaming at me
 
         // Wire up all the crap
         v.addLanguageRequestListener(p::fileChoosing);
@@ -73,14 +65,5 @@ public class TranslationHelper extends Application {
 
         p.show();
         p.fileChoosing();
-
-        Stage resxStage = new Stage(StageStyle.UNDECORATED);
-        resxStage.setScene(new Scene(rcPane));
-        resxStage.setTitle("Choose Your Resx");
-
-        primaryStage.setTitle("Translation Helper");
-        primaryStage.setScene(new Scene(overviewPane));
-        primaryStage.show();
-        resxStage.show();
     }
 }
