+++ src/main/java/de/vogel612/helper/ui/OverviewView.java
@@ -3,41 +3,28 @@ package de.vogel612.helper.ui;
 import de.vogel612.helper.data.Translation;
 
 import java.awt.event.WindowEvent;
-import java.util.HashSet;
 import java.util.List;
-import java.util.Set;
 import java.util.function.Consumer;
 
-public abstract class OverviewView {
+/**
+ * Created by vogel612 on 02.03.16.
+ */
+public interface OverviewView {
+    void addWindowClosingListener(Consumer<WindowEvent> listener);
 
-    protected final Set<Consumer<String>> translationRequestListeners = new HashSet<>();
-    protected final Set<Consumer<WindowEvent>> windowCloseListeners = new HashSet<>();
-    protected final Set<Runnable> langChoiceRequestListeners = new HashSet<>();
-    protected final Set<Runnable> saveRequestListeners = new HashSet<>();
+    void addLanguageRequestListener(Runnable listener);
 
-    public final void addWindowClosingListener(Consumer<WindowEvent> listener) {
-        windowCloseListeners.add(listener);
-    }
+    void addTranslationRequestListener(Consumer<String> listener);
 
-    public final void addLanguageRequestListener(Runnable listener) {
-        langChoiceRequestListeners.add(listener);
-    }
+    void addSaveRequestListener(Runnable listener);
 
-    public final void addTranslationRequestListener(Consumer<String> listener) {
-        translationRequestListeners.add(listener);
-    }
+    void initialize();
 
-    public final void addSaveRequestListener(Runnable listener) {
-        saveRequestListeners.add(listener);
-    }
+    void show();
 
-    public abstract void initialize();
+    void rebuildWith(List<Translation> left, List<Translation> right);
 
-    public abstract void show();
+    void displayError(String title, String errorMessage);
 
-    public abstract void rebuildWith(List<Translation> left, List<Translation> right);
-
-    public abstract void displayError(String title, String errorMessage);
-
-    public abstract void hide();
+    void hide();
 }
