+++ src/main/resources/OverviewView.fxml
@@ -5,24 +5,24 @@
 <?import javafx.scene.control.Button?>
 <?import javafx.scene.control.TableColumn?>
 <StackPane xmlns:fx="http://javafx.com/fxml"
-           fx:controller="de.vogel612.helper.ui.jfx.JFXOverviewView"
+           fx:controller="de.vogel612.helper.ui.jfx.JFXOverviewController"
            minHeight="300" minWidth="600">
-    <children>
-        <MenuBar StackPane.alignment="TOP_CENTER" StackPane.margin="5">
-            <children>
-                <Button fx:id="save" text="Save" />
-                <Button fx:id="chooseLang" text="Choose Languages" />
-            </children>
-        </MenuBar>
-        <ScrollPane StackPane.alignment="CENTER" StackPane.margin="10">
-            <children>
-                <TableView fx:id="table">
-                    <!--<columns>-->
-                        <!--<TableColumn editable="false" minWidth="250" />-->
-                        <!--<TableColumn editable="false" minWidth="250" />-->
-                    <!--</columns>-->
-                </TableView>
-            </children>
-        </ScrollPane>
-    </children>
+    <!--<children>-->
+        <!--<MenuBar StackPane.alignment="TOP_CENTER" StackPane.margin="5">-->
+            <!--<children>-->
+                <!--<Button fx:id="save" text="Save" />-->
+                <!--<Button fx:id="chooseLang" text="Choose Languages" />-->
+            <!--</children>-->
+        <!--</MenuBar>-->
+        <!--<ScrollPane StackPane.alignment="CENTER" StackPane.margin="10">-->
+            <!--<children>-->
+                <!--&lt;!&ndash;<TableView fx:id="table">&ndash;&gt;-->
+                    <!--&lt;!&ndash;<columns>&ndash;&gt;-->
+                        <!--&lt;!&ndash;<TableColumn editable="false" minWidth="250" />&ndash;&gt;-->
+                        <!--&lt;!&ndash;<TableColumn editable="false" minWidth="250" />&ndash;&gt;-->
+                    <!--&lt;!&ndash;</columns>&ndash;&gt;-->
+                <!--&lt;!&ndash;</TableView>&ndash;&gt;-->
+            <!--</children>-->
+        <!--</ScrollPane>-->
+    <!--</children>-->
 </StackPane>
