+++ src/main/resources/ResxChooser.fxml
@@ -2,7 +2,7 @@
 <?import javafx.scene.layout.*?>
 <?import javafx.geometry.Insets?>
 <GridPane xmlns:fx="http://javafx.com/fxml"
-          fx:controller="de.vogel612.helper.ui.jfx.JFXResxChooser"
+          fx:controller="de.vogel612.helper.ui.jfx.JFXResxChooserController"
           minHeight="300" minWidth="250">
     <padding>
         <Insets bottom="10" top="10" left="10" right="10" />
